+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Added :**
 
 - chapter-9.rb [examples from chapter-9]
+- one-billion-seconds.rb [todo from chapter-9]
 
 ## [0.0.11] (May 23 2021)